Update validate-working.sch
I made changes to the following checks:
 
       <!-- Check VALUE of <journal-id> element is among confirmed values --> 
Changed ‘inov” to ‘itgg’ in error message only (was correct within check)

<!-- Check <issn> EXISTS and appears in the correct number --> 
Added “dint” to the second part of the check

<!-- Check VALUE of @pub-type is correct on e-only journals --> 
Added tacl to this check

        <!-- Check only 1 <pub-date> element on e-only journals  -->  
Added tacl to this check and removed second pres 

        <!-- Check 2 <pub-date> elements on print/online journals -->  
Added tacl to this check and removed second pres 

        <!-- Check <license> and <license-p> EXIST (OA only) -->  
Added dint and itgg to this check 

        <!-- Check VALUE of @license-type @xlink:href on <license> (OA only) -->  
Restructured this check to apply to all but adev 
Created duplicate check for ADEV

  <!-- Check VALUE of <license> (OA only) --> 
Restructure this check to apply to all but adev and coli
Create two duplicates: one for adev one for coli

    <!-- Check VALUE of <copyright-holder> (OA only: COLI)--> 

Remove tacl from this
Create two duplicates: tacl, dint
NOTES: realized this check wasn’t working properly and I fixed the error in all copyright holder checks. For some reason the “report” structure was not working, so I changed to an “assert” structure. 

    <!-- Check VALUE of <copyright-statement> (OA only: COLI)-->
Remove tacl from this
Create two duplicates: tacl, dint

    <!-- Check VALUE of <corresp> --> 
Removed this check. This isn’t consistent across all journals because of necessity, but should be something a PC can check visually. If it makes sense it is OK.

        <!-- Check VALUE of <contrib-id> --> 
added missing “s” in message 

        <!-- Check <abstract> DOES NOT contain <title> -->
Adjusted to allow netn to have a title in their abstract, because this is what they are doing.